+++ src/lib/ProjectLinks.svelte
@@ -19,16 +19,20 @@
 	{/each}
 </menu>
 <small class="display:block text-align:right">
-	the charts show public commit activity<br />since Gro planted on August 7, 2019
+	the charts show public commit activity<br />since Gro started in August 2019 -<br />most of the
+	repos created in 2023-2024 are<br />history-resetting forks from 2020~2023 attempts
 </small>
 
 {#snippet package_thumbnail(project_item: ProjectItem)}
 	{@const counts = series_by_name.get(project_item.name)}
-	<!-- eslint-disable-next-line svelte/no-navigation-without-resolve -->
-	<a class="thumbnail row shade_00 px_md py_xs border_radius_sm mb_lg" href={project_item.url}>
+	<!-- eslint-disable svelte/no-navigation-without-resolve -->
+	<a
+		class="thumbnail row align-items:start shade_00 px_md py_xs border_radius_sm mb_lg"
+		href={project_item.url}
+	>
 		<Svg shrink={false} data={project_item.logo} size="var(--icon_size_lg)" />
 		<div class="pl_lg width_atmost_sm">
-			<div class="thumbnail-name row align-items:start {project_item.color_class}">
+			<div class="thumbnail-name row {project_item.color_class}">
 				{project_item.name}
 				{#if counts}
 					{@const stats_label = label_by_name.get(project_item.name)}
@@ -43,6 +47,7 @@
 			</div>
 		</div>
 	</a>
+	<!-- eslint-enable svelte/no-navigation-without-resolve -->
 {/snippet}
 
 <style>